Fix zlib uncompression
The standard library throws when zlib returns Z_BUF_ERROR.
However as per official documentation this is not an error.
Rather is states that there is no more input being written.
This for example happens when the compressed size is equal to the
uncompressed size.

This commit treats Z_BUF_ERROR the same as Z_STREAM_END.