getFileData has a new override and new parameter
getFileData can now be called with an explicit filehandle
instead of the one provided by the GRF struct. This should
enable multithreading the extraction process.

It is also now possible to provide a cache flag. If set to
No.useCache then the data won't be stored in the file structs
data member. If set to Yes.useCache then it will be and following
calls when return file.data directly instead of processing the data
once more.